requirements.txt
pipecat-ai[webrtc,deepgram,groq,silero,runner]
python-dotenv
aiosqlite
loguru
